Remove unused imports and typecast in keypress loop
The imports for `foundation.hpp` and `gfx/platform_surface.hpp` were removed as they were not being used. Additionally, the type of `i` in the loop for checking pressed keys was changed from size_t to int, and the loop was adjusted accordingly.

+++ src/extra/gfx/main_window.cpp
@@ -9,8 +9,6 @@
 #include <gfx/window.hpp>
 #include "../inputs.hpp"
 #include "extra/gfx.hpp"
-#include "foundation.hpp"
-#include "gfx/platform_surface.hpp"
 #include "linalg_shim.hpp"
 #include "params.hpp"
 
@@ -225,7 +223,7 @@ struct MainWindow final {
     int numKeys{};
     auto keyStates = SDL_GetKeyboardState(&numKeys);
     _actuallyHeldKeys.clear();
-    for (size_t i = 0; i < numKeys; i++) {
+    for (int i = 0; i < numKeys; i++) {
       SDL_Keycode code = SDL_GetKeyFromScancode((SDL_Scancode)i);
       if (keyStates[i] == 1) {
         _actuallyHeldKeys.insert(code);
